Moved circularly-referenced files from engine to core
