+++ builtin/js/mldb_js.cc
@@ -15,8 +15,8 @@
 #include "mldb/core/function.h"
 #include "mldb/core/sensor.h"
 #include "mldb/base/optimized_path.h"
-#include "mldb/engine/dataset_scope.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/dataset_scope.h"
+#include "mldb/core/analytics.h"
 #include "procedure_js.h"
 #include "function_js.h"
 #include "sensor_js.h"
