+++ builtin/matrix.cc
@@ -10,9 +10,9 @@
 #include "mldb/types/basic_value_descriptions.h"
 #include "mldb/base/parallel.h"
 #include "mldb/arch/timers.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/sql/sql_expression.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/types/annotated_exception.h"
 #include "mldb/base/less.h"
 #include "mldb/utils/log.h"
