+++ builtin/pooling_function.cc
@@ -8,8 +8,8 @@
 #include "mldb/core/mldb_engine.h"
 #include "mldb/types/basic_value_descriptions.h"
 #include "mldb/types/any_impl.h"
-#include "mldb/engine/dataset_scope.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/dataset_scope.h"
+#include "mldb/core/analytics.h"
 #include "mldb/core/dataset.h"
 #include "mldb/utils/string_functions.h"
 #include "mldb/utils/profile.h"
