+++ builtin/sql_functions.cc
@@ -8,11 +8,11 @@
 #include "mldb/builtin/sql_functions.h"
 #include "mldb/core/mldb_engine.h"
 #include "mldb/sql/sql_expression.h"
-#include "mldb/engine/dataset_scope.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/types/basic_value_descriptions.h"
 #include "mldb/base/parallel.h"
-#include "mldb/engine/bound_queries.h"
+#include "mldb/core/bound_queries.h"
 #include "mldb/sql/table_expression_operations.h"
 #include "mldb/sql/join_utils.h"
 #include "mldb/sql/execution_pipeline.h"
@@ -21,7 +21,7 @@
 #include "mldb/base/per_thread_accumulator.h"
 #include "mldb/rest/in_process_rest_connection.h"
 #include "mldb/builtin/sql_config_validator.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/utils/log.h"
 #include "mldb/rest/cancellation_exception.h"
 #include <memory>
