+++ builtin/sub_dataset.cc
@@ -6,8 +6,8 @@
 */
 
 #include "mldb/sql/sql_expression.h"
-#include "mldb/engine/dataset_scope.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/dataset_scope.h"
+#include "mldb/core/analytics.h"
 #include "sub_dataset.h"
 #include "mldb/types/any_impl.h"
 #include "mldb/utils/lightweight_hash.h"
