+++ core/analytics.h
