+++ core/dataset.cc
@@ -14,14 +14,14 @@
 #include "mldb/sql/sql_expression_operations.h"
 #include "mldb/types/tuple_description.h"
 #include "mldb/types/vector_description.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/sql/sql_expression.h"
 #include "mldb/sql/sql_utils.h"
 #include "mldb/utils/lightweight_hash.h"
 #include "mldb/utils/profile.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/base/per_thread_accumulator.h"
-#include "mldb/engine/bucket.h"
+#include "mldb/core/bucket.h"
 #include "mldb/base/parallel_merge_sort.h"
 #include "mldb/utils/environment.h"
 #include "mldb/base/parallel.h"
