+++ engine/column_scope.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 
 
 namespace MLDB {
