+++ plugins/feature_gen/bucketize_procedure.cc
@@ -7,10 +7,10 @@
 #include "bucketize_procedure.h"
 #include "mldb/core/mldb_engine.h"
 #include "mldb/sql/sql_expression.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/types/basic_value_descriptions.h"
 #include "mldb/base/parallel.h"
-#include "mldb/engine/bound_queries.h"
+#include "mldb/core/bound_queries.h"
 #include "mldb/sql/table_expression_operations.h"
 #include "mldb/sql/join_utils.h"
 #include "mldb/sql/execution_pipeline.h"
