+++ plugins/feature_gen/dist_table_procedure.cc
@@ -13,14 +13,14 @@
 #include "mldb/types/vector_description.h"
 #include "mldb/types/map_description.h"
 #include "mldb/rest/in_process_rest_connection.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/builtin/matrix.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "types/any_impl.h"
 #include "mldb/utils/string_functions.h"
 #include "mldb/builtin/sql_functions.h"
 #include "sql/execution_pipeline.h"
-#include "mldb/engine/bound_queries.h"
+#include "mldb/core/bound_queries.h"
 #include "mldb/types/annotated_exception.h"
 #include "mldb/types/db/persistent.h"
 #include "mldb/types/jml_serialization.h"
