+++ plugins/jml/classifier.cc
@@ -12,7 +12,7 @@
 #include "dataset_feature_space.h"
 #include "mldb/core/mldb_engine.h"
 #include "mldb/core/dataset.h"
-#include "mldb/engine/bound_queries.h"
+#include "mldb/core/bound_queries.h"
 #include "mldb/sql/sql_expression.h"
 #include "mldb/utils/distribution.h"
 #include "mldb/base/scope.h"
@@ -36,7 +36,7 @@
 #include "mldb/plugins/jml/jml/onevsall_generator.h"
 #include "mldb/base/map_reduce.h"
 #include "mldb/utils/string_functions.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/base/per_thread_accumulator.h"
 #include "mldb/base/parallel_merge_sort.h"
 #include "mldb/plugins/jml/jml/feature_info.h"
