+++ plugins/jml/em_interface.cc
@@ -22,7 +22,7 @@
 #include "mldb/utils/vector_utils.h"
 #include "mldb/types/basic_value_descriptions.h"
 #include "mldb/sql/sql_expression.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/types/any_impl.h"
 #include "mldb/utils/smart_ptr_utils.h"
 #include "mldb/vfs/fs_utils.h"
