+++ plugins/jml/experiment_procedure.cc
@@ -10,9 +10,9 @@
 #include "types/distribution_description.h"
 #include "types/map_description.h"
 #include "mldb/rest/in_process_rest_connection.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/builtin/matrix.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/types/any_impl.h"
 #include "mldb/utils/string_functions.h"
 #include "arch/timers.h"
