+++ plugins/jml/kmeans_interface.cc
@@ -23,7 +23,7 @@
 #include "mldb/plugins/jml/kmeans.h"
 #include "mldb/sql/sql_expression.h"
 #include "mldb/builtin/sql_config_validator.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/types/any_impl.h"
 #include "mldb/types/optional_description.h"
 #include "mldb/vfs/fs_utils.h"
