+++ plugins/pro/list_files_procedure.cc
@@ -8,9 +8,9 @@
 #include "mldb/vfs/fs_utils.h"
 #include "mldb/core/mldb_engine.h"
 #include "mldb/sql/sql_expression.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/types/basic_value_descriptions.h"
-#include "mldb/engine/bound_queries.h"
+#include "mldb/core/bound_queries.h"
 #include "mldb/types/any_impl.h"
 #include "mldb/types/date.h"
 #include "mldb/plugins/pro/pro_plugin.h"
