+++ plugins/tensorflow/tensorflow_plugin.cc
@@ -21,7 +21,7 @@
 #include "mldb/utils/string_functions.h"
 #include "mldb/vfs/filter_streams.h"
 #include "mldb/vfs/fs_utils.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/rest/rest_request_binding.h"
 #include "mldb/engine/static_content_macro.h"
 #include "mldb/sql/builtin_functions.h"
