+++ plugins/textual/importtext_procedure.cc
@@ -17,7 +17,7 @@
 #include "mldb/sql/sql_expression.h"
 #include "mldb/types/basic_value_descriptions.h"
 #include "mldb/types/any_impl.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/vfs/filter_streams.h"
 #include "mldb/vfs/fs_utils.h"
 #include "mldb/utils/progress.h"
