+++ server/mldb_server.cc
@@ -30,10 +30,10 @@
 #include "mldb/engine/credential_collection.h"
 #include "mldb/engine/sensor_collection.h"
 #include "mldb/engine/procedure_run_collection.h"
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/vfs/fs_utils.h"
 #include "mldb/vfs/filter_streams.h"
-#include "mldb/engine/analytics.h"
+#include "mldb/core/analytics.h"
 #include "mldb/types/meta_value_description.h"
 #include "mldb/arch/simd.h"
 #include "mldb/utils/log.h"
