+++ sql/sql_expression_operations.cc
@@ -14,7 +14,7 @@
 #include "mldb/types/set_description.h"
 #include "table_expression_operations.h"
 #include <unordered_set>
-#include "mldb/engine/dataset_scope.h"
+#include "mldb/core/dataset_scope.h"
 #include "mldb/base/scope.h"
 #include "mldb/sql/sql_utils.h"
 #include "mldb/utils/distribution.h"
